+++ chain/core/Cargo.toml
@@ -35,3 +35,4 @@ features = ["derive"]
 
 [dev-dependencies]
 serde_json = "1.0"
+hex-literal = "1.0"
